+++ lib/Handler/TextDocument/TextDocumentHandler.php
@@ -11,7 +11,7 @@ use Phpactor\LanguageServerProtocol\TextDocumentSyncKind;
 use Phpactor\LanguageServerProtocol\WillSaveTextDocumentParams;
 use Phpactor\LanguageServer\Core\Handler\CanRegisterCapabilities;
 use Phpactor\LanguageServer\Core\Handler\Handler;
-use Phpactor\LanguageServer\Core\Session\Workspace;
+use Phpactor\LanguageServer\Core\Workspace\Workspace;
 use Phpactor\LanguageServer\Event\TextDocumentClosed;
 use Phpactor\LanguageServer\Event\TextDocumentOpened;
 use Phpactor\LanguageServer\Event\TextDocumentSaved;
